requirements.txt
streamlit>=1.31.0
fastapi>=0.109.0
uvicorn>=0.27.0
langchain>=0.1.0
langchain-community>=0.0.16
openai>=1.12.0
python-dotenv>=1.0.0
pydantic>=2.6.0
requests>=2.31.0
pillow>=10.2.0
pandas>=2.2.0
python-multipart>=0.0.6 